WEEK-2/MAY-14/requirements.txt
streamlit
google_generativeai
python-dotenv
langchain
chromadb
PyMuPDF
langchain_google_genai